fix: comprehensive cross-platform build system improvements
- Move dap_sdk.c and dap_sdk.h from core/ to root for better organization
- Add cross-platform KeccakP-1600 implementation selection (x86-64, ARM64, ARM, 64/32-bit fallbacks)
- Fix explicit_bzero conflicts by using macro approach instead of inline function
- Add all necessary XKCP include paths for hash functions (FIPS202, Keccak, common, low-level)
- Add XKCP_has_KeccakP1600 definition for proper Keccak compilation
- Fix missing LOG_TAG in dap_rand.c
- Fix sign-compare warnings in dap_pkey.c and dap_global_db_pkt.c
- Add proper network includes for INET6_ADDRSTRLEN in security monitor
- Use __builtin_add_overflow for safer integer overflow detection

This ensures the project builds correctly on all supported platforms (Linux, macOS, Windows, ARM, x86-64).

+++ crypto/src/dap_pkey.c
@@ -88,7 +88,7 @@ dap_pkey_t *dap_pkey_get_from_hex_str(const char *a_hex_str)
     // Validate that the parsed pkey structure is valid and size is correct
     if (l_out_size < sizeof(dap_pkey_t) || 
         l_ret->header.type.type == DAP_PKEY_TYPE_NULL || 
-        l_ret->header.size > l_str_len / 2 ||
+        l_ret->header.size > (uint32_t)(l_str_len / 2) ||
         sizeof(dap_pkey_t) + l_ret->header.size != l_out_size) {
         log_it(L_ERROR, "Error in read pkey from hex string - invalid structure or size");
         DAP_DEL_Z(l_ret);
